Extract new base test class KernelTestClass
Ideal base class that needs access to a Kernel but not a Client such as
when testing a Command.
